+++ instrumentation/aws-sdk/aws-sdk-1.11/library/src/main/java/io/opentelemetry/instrumentation/awssdk/v1_11/internal/SqsAccess.java
@@ -5,12 +5,15 @@
 
 package io.opentelemetry.instrumentation.awssdk.v1_11.internal;
 
+import static java.util.Collections.emptyList;
+
 import com.amazonaws.AmazonWebServiceRequest;
 import com.amazonaws.Request;
 import com.amazonaws.Response;
 import io.opentelemetry.context.Context;
 import io.opentelemetry.instrumentation.api.internal.Timer;
 import io.opentelemetry.javaagent.tooling.muzzle.NoMuzzle;
+import java.util.Collection;
 import javax.annotation.Nullable;
 
 final class SqsAccess {
@@ -42,6 +45,11 @@ final class SqsAccess {
   @Nullable
   static String getMessageAttribute(Request<?> request, String name) {
     return enabled ? SqsImpl.getMessageAttribute(request, name) : null;
+  }
+
+  @NoMuzzle
+  static Collection<String> getMessageAttributeNames(Request<?> request) {
+    return enabled ? SqsImpl.getMessageAttributeNames(request) : emptyList();
   }
 
   @NoMuzzle