+++ instrumentation/aws-sdk/aws-sdk-1.11/library/src/main/java/io/opentelemetry/instrumentation/awssdk/v1_11/internal/SqsMessage.java
@@ -6,6 +6,7 @@
 package io.opentelemetry.instrumentation.awssdk.v1_11.internal;
 
 import io.opentelemetry.context.Context;
+import java.util.Collection;
 import java.util.Map;
 import javax.annotation.Nullable;
 
@@ -21,6 +22,8 @@ interface SqsMessage {
 
   @Nullable
   String getMessageAttribute(String name);
+
+  Collection<String> getMessageAttributeNames();
 
   String getMessageId();
 }