+++ instrumentation/nats/nats-2.17/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/nats/v2_17/NatsSingletons.java
@@ -10,24 +10,23 @@ import static io.opentelemetry.instrumentation.nats.v2_17.internal.NatsInstrumen
 import static io.opentelemetry.instrumentation.nats.v2_17.internal.NatsInstrumenterFactory.createRequestInstrumenter;
 
 import io.opentelemetry.api.GlobalOpenTelemetry;
+import io.opentelemetry.instrumentation.api.config.IncludeExclude;
 import io.opentelemetry.instrumentation.api.instrumenter.Instrumenter;
 import io.opentelemetry.instrumentation.nats.v2_17.internal.NatsRequest;
 import io.opentelemetry.javaagent.bootstrap.internal.ExperimentalConfig;
-import java.util.List;
 
 class NatsSingletons {
 
-  private static final List<String> capturedHeaders =
-      ExperimentalConfig.get().getMessagingHeaders();
+  private static final IncludeExclude headers = ExperimentalConfig.get().getMessagingHeaders();
 
   private static final Instrumenter<NatsRequest, NatsRequest> publishInstrumenter =
-      createPublishInstrumenter(GlobalOpenTelemetry.get(), capturedHeaders);
+      createPublishInstrumenter(GlobalOpenTelemetry.get(), headers);
 
   private static final Instrumenter<NatsRequest, NatsRequest> requestInstrumenter =
-      createRequestInstrumenter(GlobalOpenTelemetry.get(), capturedHeaders);
+      createRequestInstrumenter(GlobalOpenTelemetry.get(), headers);
 
   private static final Instrumenter<NatsRequest, Void> consumerProcessInstrumenter =
-      createConsumerProcessInstrumenter(GlobalOpenTelemetry.get(), capturedHeaders);
+      createConsumerProcessInstrumenter(GlobalOpenTelemetry.get(), headers);
 
   static Instrumenter<NatsRequest, NatsRequest> publishInstrumenter() {
     return publishInstrumenter;