+++ instrumentation/nats/nats-2.17/library/src/main/java/io/opentelemetry/instrumentation/nats/v2_17/internal/NatsInstrumenterFactory.java
@@ -9,13 +9,14 @@ import static io.opentelemetry.instrumentation.api.incubator.semconv.messaging.i
 import static io.opentelemetry.instrumentation.api.incubator.semconv.messaging.internal.MessagingExceptionEventExtractors.setMessagingSendExceptionEventExtractor;
 
 import io.opentelemetry.api.OpenTelemetry;
+import io.opentelemetry.instrumentation.api.config.IncludeExclude;
 import io.opentelemetry.instrumentation.api.incubator.semconv.messaging.MessagingAttributesExtractor;
 import io.opentelemetry.instrumentation.api.incubator.semconv.messaging.MessagingOperationType;
 import io.opentelemetry.instrumentation.api.incubator.semconv.messaging.MessagingSpanNameExtractor;
 import io.opentelemetry.instrumentation.api.incubator.semconv.messaging.internal.MessagingProcessInstrumenterFactory;
+import io.opentelemetry.instrumentation.api.instrumenter.AttributesExtractor;
 import io.opentelemetry.instrumentation.api.instrumenter.Instrumenter;
 import io.opentelemetry.instrumentation.api.instrumenter.InstrumenterBuilder;
-import java.util.List;
 
 /**
  * This class is internal and is hence not for public use. Its APIs are unstable and can change at
@@ -30,17 +31,17 @@ public final class NatsInstrumenterFactory {
   private static final String PROCESS_OPERATION_NAME = "process";
 
   public static Instrumenter<NatsRequest, NatsRequest> createPublishInstrumenter(
-      OpenTelemetry openTelemetry, List<String> capturedHeaders) {
-    return createProducerInstrumenter(openTelemetry, capturedHeaders, PUBLISH_OPERATION_NAME);
+      OpenTelemetry openTelemetry, IncludeExclude headers) {
+    return createProducerInstrumenter(openTelemetry, headers, PUBLISH_OPERATION_NAME);
   }
 
   public static Instrumenter<NatsRequest, NatsRequest> createRequestInstrumenter(
-      OpenTelemetry openTelemetry, List<String> capturedHeaders) {
-    return createProducerInstrumenter(openTelemetry, capturedHeaders, REQUEST_OPERATION_NAME);
+      OpenTelemetry openTelemetry, IncludeExclude headers) {
+    return createProducerInstrumenter(openTelemetry, headers, REQUEST_OPERATION_NAME);
   }
 
   private static Instrumenter<NatsRequest, NatsRequest> createProducerInstrumenter(
-      OpenTelemetry openTelemetry, List<String> capturedHeaders, String operationName) {
+      OpenTelemetry openTelemetry, IncludeExclude headers, String operationName) {
     InstrumenterBuilder<NatsRequest, NatsRequest> builder =
         Instrumenter.<NatsRequest, NatsRequest>builder(
                 openTelemetry,
@@ -50,18 +51,13 @@ public final class NatsInstrumenterFactory {
                     MessagingOperationType.SEND,
                     operationName))
             .addAttributesExtractor(
-                MessagingAttributesExtractor.builder(
-                        new NatsRequestMessagingAttributesGetter(),
-                        MessagingOperationType.SEND,
-                        operationName)
-                    .setCapturedHeaders(capturedHeaders)
-                    .build());
+                messagingAttributesExtractor(MessagingOperationType.SEND, operationName, headers));
     setMessagingSendExceptionEventExtractor(builder);
     return builder.buildProducerInstrumenter(new NatsRequestTextMapSetter());
   }
 
   public static Instrumenter<NatsRequest, Void> createConsumerProcessInstrumenter(
-      OpenTelemetry openTelemetry, List<String> capturedHeaders) {
+      OpenTelemetry openTelemetry, IncludeExclude headers) {
     InstrumenterBuilder<NatsRequest, Void> builder =
         Instrumenter.<NatsRequest, Void>builder(
                 openTelemetry,
@@ -71,12 +67,8 @@ public final class NatsInstrumenterFactory {
                     MessagingOperationType.PROCESS,
                     PROCESS_OPERATION_NAME))
             .addAttributesExtractor(
-                MessagingAttributesExtractor.builder(
-                        new NatsRequestMessagingAttributesGetter(),
-                        MessagingOperationType.PROCESS,
-                        PROCESS_OPERATION_NAME)
-                    .setCapturedHeaders(capturedHeaders)
-                    .build());
+                messagingAttributesExtractor(
+                    MessagingOperationType.PROCESS, PROCESS_OPERATION_NAME, headers));
     setMessagingProcessExceptionEventExtractor(builder);
 
     return MessagingProcessInstrumenterFactory.create(
@@ -84,6 +76,14 @@ public final class NatsInstrumenterFactory {
         openTelemetry.getPropagators().getTextMapPropagator(),
         new NatsRequestTextMapGetter(),
         false);
+  }
+
+  private static AttributesExtractor<NatsRequest, Object> messagingAttributesExtractor(
+      MessagingOperationType operationType, String operationName, IncludeExclude headers) {
+    return MessagingAttributesExtractor.builder(
+            new NatsRequestMessagingAttributesGetter(), operationType, operationName)
+        .setHeaders(headers)
+        .build();
   }
 
   private NatsInstrumenterFactory() {}